refactor(coverage-planning): move section components to controls-panel folder
