+++ src/features/coverage-planning/components/controls-panel/AlgoSection/PathWidthInput/PathWidthInput.tsx
