+++ src/features/coverage-planning/components/controls-panel/EnvSection/FormatSelection/FormatSelection.tsx
