+++ src/features/coverage-planning/components/controls-panel/EnvSection/GlobalTypeCheckbox/GlobalTypeCheckbox.tsx
