+++ src/features/coverage-planning/components/controls-panel/EnvSection/TypeSelection/TypeSelection.tsx
